requirements.txt
python-socketio
pyautogui
aiohttp